+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Modified
 - responsive navigation
 
+### Fixed
+- fix static management page
+
 ## [0.8.7] - 2019-11-24
 ### Added
 - description for guilds